댓글 조회 구현

+++ src/main/java/com/sparta/schedulemanagement_spring/controller/CommentsController.java
@@ -6,6 +6,8 @@ import com.sparta.schedulemanagement_spring.service.CommentsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequestMapping("/api")
 @RequiredArgsConstructor
@@ -18,19 +20,20 @@ public class CommentsController {
         return commentsService.addComment(scheduleId,requestDto);
     }
 
-   /* // 댓글 조회
+   // 댓글 조회
     @GetMapping("/schedules/{scheduleId}/comments")
-    public List<CommentsResponseDto> getAllComments(@PathVariable int scheduleId) {
+    public List<CommentsResponseDto> getAllComments(@PathVariable Long scheduleId) {
         return commentsService.getAllComments(scheduleId);
     }
+    /*
     // 댓글 수정
     @PutMapping("/schedules/{scheduleId}/comments/{commentId}")
-    public CommentsResponseDto updateComment(@PathVariable int scheduleId, @PathVariable int commentId, @RequestBody CommentsRequestDto requestDto) {
+    public CommentsResponseDto updateComment(@PathVariable Long scheduleId, @PathVariable Long commentId, @RequestBody CommentsRequestDto requestDto) {
         return commentsService.updateComment(scheduleId,commentId,requestDto);
     }
     // 댓글 삭제
     @DeleteMapping("/schedules/{scheduleId}/comments/{commentId}")
-    public CommentsResponseDto deleteComment(@PathVariable int scheduleId, @PathVariable int commentId) {
+    public CommentsResponseDto deleteComment(@PathVariable Long scheduleId, @PathVariable Long commentId) {
         return commentsService.deleteComment(scheduleId,commentId);
     }*/
